Move shim signal handling to another file
