+++ src/lib/shim/CMakeLists.txt
@@ -37,6 +37,7 @@ set(SHIM_FILES
   shim_logger.c
   shim_rdtsc.c
   shim_seccomp.c
+  shim_signals.c
   shim_sys.c
   shim_syscall.c
   shim_tls.c
